Drop forecast toggle
