+++ plugins/CoreHome/lang/en.json
@@ -32,10 +32,8 @@
         "FlattenDataTable": "The report is hierarchical %s Make it flat",
         "FormatMetrics": "Format metrics",
         "HideExportUrl": "Hide Export URL",
-        "HideForecast": "Hide forecast",
         "HomeShortcut": "Home",
         "ShowExportUrl": "Show Export URL",
-        "ShowForecast": "Show forecast",
         "SupportUsOn": "Support us on",
         "IncludeRowsWithLowPopulation": "Rows with low population are hidden %s Show all rows",
         "InjectedHostEmailBody": "Hello, I tried to access Matomo today and encountered the unknown hostname warning.",
