+++ plugins/CoreVisualizations/Visualizations/JqplotGraph/Evolution/Config.php
@@ -25,17 +25,21 @@ class Config extends JqplotGraphConfig
     public $show_line_graph = true;
 
     /**
-     * Whether forecast values should be rendered for incomplete periods.
+     * Whether forecast values should be rendered for incomplete periods. The forecast is
+     * always active; whether anything is actually drawn still depends on feasibility (line
+     * chart, not {@see $disable_forecast}, and at least one incomplete period that yields a
+     * renderable value). Forced off in bar mode by the visualization's beforeLoadDataTable()
+     * since the BarRenderer has nowhere to draw forecast points.
      *
-     * Default value: false
+     * Default value: true
      */
-    public $show_forecast = false;
+    public $show_forecast = true;
 
     /**
      * Hard gate that suppresses the forecast feature regardless of {@see $show_forecast}.
-     * Hides the forecast toggle action and skips the precompute path so callers that fan
-     * out into label-filtered inner API calls (e.g. row evolution popovers) do not pay for
-     * the sub-period blob fetches the forecast builder consumes.
+     * Skips the precompute path so callers that fan out into label-filtered inner API calls
+     * (e.g. row evolution popovers) do not pay for the sub-period blob fetches the forecast
+     * builder consumes.
      *
      * Default value: false
      */
@@ -51,11 +55,11 @@ class Config extends JqplotGraphConfig
         $this->hide_annotations_view  = false;
         $this->x_axis_step_size       = false;
         $this->show_line_graph        = true;
-        $this->show_forecast          = false;
+        $this->show_forecast          = true;
         $this->disable_forecast       = false;
 
-        $this->addPropertiesThatShouldBeAvailableClientSide(['show_line_graph', 'show_forecast']);
-        $this->addPropertiesThatCanBeOverwrittenByQueryParams(['show_line_graph', 'show_forecast']);
+        $this->addPropertiesThatShouldBeAvailableClientSide(['show_line_graph']);
+        $this->addPropertiesThatCanBeOverwrittenByQueryParams(['show_line_graph']);
 
         $period = Common::getRequestVar('period');
         if ($period !== 'range') {
